split task into workflow and task to make the strcture more clear
